refactor: cache package
